Add IPC::Run3 to register test file

+++ t/55-register.t
@@ -2,9 +2,12 @@ use warnings;
 use strict;
 
 use lib 't/';
+
 use BB;
+use Capture::Tiny qw(:all);
 use File::Copy;
 use File::Path qw(make_path);
+use IPC::Run3;
 use Test::More;
 use Win32::TieRegistry;
 
@@ -22,21 +25,23 @@ if (! -d $dir){
 
 unlink $customfile or die $! if -f $customfile;
 
-my $o;
+my ($o, $err);
 
-mkdir "$dir/empty" or die $!;
+if (! -d "$dir/empty") {
+    mkdir "$dir/empty" or die $!;
+}    
 is -d "$dir/empty", 1, "created empty installation dir ok";
 
-$o = `$c register empty`;
-like $o, qr/empty is not a valid Perl installation/, "no registration if a perl binary not found";
-
+$err = capture_stderr { eval { run3 "$c register empty"; }; };
+is $? >> 8, BB::err_code('PERL_INVALID_ERROR'), "registration failure if perl binary not found sets exit status ok";
+like $err, qr/empty is not a valid Perl installation/, "no registration if a perl binary not found error msg ok";
 
 rmdir "$dir/empty" or die $!;
 is -d "$dir/empty", undef, "removed empty instance ok";
 
-$o = `$c register not_exist`;
-like $o, qr/installation directory.*does not exist/, "won't register if dir doesn't exist ok";
-
+$err = capture_stderr { eval { run3 "$c register not_exist"; }; };
+is $? >> 8, BB::err_code('DIRECTORY_NOT_EXIST'), "register failure if perl dir no exist sets exit status ok";
+like $err, qr/installation directory.*does not exist/, "won't register if dir doesn't exist errmsg ok";
 
 my @avail = BB::get_avail();
 my @installed = BB::get_installed();
@@ -76,10 +81,11 @@ is -f "$dir/dup/perl/bin/perl.exe", 1, "test 'dup' directory created ok";
 `$c register dup`;
 
 $o = `$c available`;
-like $o, qr/dup.*\[custom/, "registered a valid instance ok";
+like $o, qr/dup.*\[custom/, "registered a valid instance (dup) ok";
 
-$o = `$c register dup`;
-like $o, qr/dup instance is already registered/, "don't duplicate registration ok";
+$err = capture_stderr { eval { run3 "$c register dup"; }; };
+is $? >> 8, BB::err_code('PERL_VERSION_ALREADY_REGISTERED'), "exit status for already registered Perl ok";
+like $err, qr/dup instance is already registered/, "don't duplicate registration errmsg ok";
 
 for (BB::get_installed()){
     $o = `$c remove $_`;
